Removed LLVMTypeGen
